fix: split component file
