+++ packages/react/components/Toast/Hook.ts
@@ -1,0 +1,9 @@
+import { addToast, update, close } from "./Store";
+
+export function useToast() {
+  return {
+    toast: addToast,
+    update,
+    close,
+  };
+}
